Clippy fixes.

+++ src/ui/util.rs
@@ -32,5 +32,5 @@ pub const fn get_usage_color(usage: u16, regular_color: Color) -> Color {
 }
 
 pub fn get_spark_line_len() -> u16 {
-    get_terminal_width() as u16 / 2 - config::PRETTY_BYTES_COL_LEN as u16
+    get_terminal_width() as u16 / 2 - config::PRETTY_BYTES_COL_LEN
 }
